only throw empty exception in refresh method

+++ app/src/main/java/com/github/hongbeomi/flickrcodelab/data/source/DefaultPhotoListRepository.kt
@@ -1,9 +1,11 @@
 package com.github.hongbeomi.flickrcodelab.data.source
 
+import com.github.hongbeomi.flickrcodelab.data.source.remote.EXCEPTION_MESSAGE_LIST_EMPTY
 import com.github.hongbeomi.flickrcodelab.model.Photo
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.single
 import kotlinx.coroutines.withContext
 
 class DefaultPhotoListRepository(
@@ -22,6 +24,10 @@ class DefaultPhotoListRepository(
 
     override suspend fun refreshPhotoList() = withContext(dispatcher) {
         val newPhotoList = remotePhotosDataSource.getSearchPhotoList()
+
+        if (newPhotoList.single().isEmpty()) {
+            throw IllegalStateException(EXCEPTION_MESSAGE_LIST_EMPTY)
+        }
 
         newPhotoList.collect {
             localPhotosDataSource.deleteAllPhotoList()
